Move Bargraph functionality into /view
